backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-genai==1.52.0
python-dotenv==1.0.1
fpdf2==2.7.9
Pillow==10.4.0
PyPDF2==3.0.1
langchain==1.0.8
langchain-community==0.4.1
langchain-google-genai==3.1.0
langchain-text-splitters==1.0.0
faiss-cpu==1.8.0.post1
pytest==8.3.3
